Fix bg colour in cf endpoint card given error - includes fixes to cf/endpoint list data sources inheritance -
