+++ src/frontend/packages/core/src/shared/components/list/list-types/cf-endpoints/cf-endpoints-data-source.ts
@@ -1,13 +1,14 @@
 import { Store } from '@ngrx/store';
 
-import { ListDataSource } from '../../data-sources-controllers/list-data-source';
-import { IListConfig } from '../../list.component.types';
-import { AppState } from '../../../../../../../store/src/app-state';
 import { GetAllEndpoints } from '../../../../../../../store/src/actions/endpoint.actions';
 import { CreatePagination } from '../../../../../../../store/src/actions/pagination.actions';
+import { AppState } from '../../../../../../../store/src/app-state';
 import { EndpointModel } from '../../../../../../../store/src/types/endpoint.types';
-import { entityFactory, endpointSchemaKey } from '../../../../../../../store/src/helpers/entity-factory';
-import { GetSystemInfo } from '../../../../../../../store/src/actions/system.actions';
+import { EntityMonitorFactory } from '../../../../monitors/entity-monitor.factory.service';
+import { InternalEventMonitorFactory } from '../../../../monitors/internal-event-monitor.factory';
+import { PaginationMonitorFactory } from '../../../../monitors/pagination-monitor.factory';
+import { IListConfig } from '../../list.component.types';
+import { BaseEndpointsDataSource } from '../endpoint/base-endpoints-data-source';
 
 function syncPaginationSection(
   store: Store<AppState>,
@@ -20,38 +21,21 @@ function syncPaginationSection(
     action.paginationKey
   ));
 }
-export class CFEndpointsDataSource extends ListDataSource<EndpointModel> {
+export class CFEndpointsDataSource extends BaseEndpointsDataSource {
   store: Store<AppState>;
 
   constructor(
     store: Store<AppState>,
-    listConfig: IListConfig<EndpointModel>
+    listConfig: IListConfig<EndpointModel>,
+    paginationMonitorFactory: PaginationMonitorFactory,
+    entityMonitorFactory: EntityMonitorFactory,
+    internalEventMonitorFactory: InternalEventMonitorFactory
   ) {
     const action = new GetAllEndpoints();
     const paginationKey = 'cf-endpoints';
     // We do this here to ensure we sync up with main endpoint table data.
     syncPaginationSection(store, action, paginationKey);
     action.paginationKey = paginationKey;
-    super({
-      store,
-      action,
-      schema: entityFactory(endpointSchemaKey),
-      getRowUniqueId: object => object.guid,
-      paginationKey,
-      isLocal: true,
-      transformEntities: [
-        (entities: EndpointModel[]) => {
-          return entities.filter(endpoint => {
-            return endpoint.connectionStatus === 'connected' && endpoint.cnsi_type === 'cf';
-          });
-        },
-        {
-          type: 'filter',
-          field: 'name'
-        },
-      ],
-      listConfig,
-      refresh: () => this.store.dispatch(new GetSystemInfo(false, action))
-    });
+    super(store, listConfig, action, 'cf', paginationMonitorFactory, entityMonitorFactory, internalEventMonitorFactory);
   }
 }
