+++ src/frontend/packages/core/src/shared/components/list/list-types/cf-endpoints/cf-endpoints-list-config.service.ts
@@ -3,18 +3,21 @@ import { Store } from '@ngrx/store';
 
 import { AppState } from '../../../../../../../store/src/app-state';
 import { EndpointModel } from '../../../../../../../store/src/types/endpoint.types';
+import { EntityMonitorFactory } from '../../../../monitors/entity-monitor.factory.service';
+import { InternalEventMonitorFactory } from '../../../../monitors/internal-event-monitor.factory';
+import { PaginationMonitorFactory } from '../../../../monitors/pagination-monitor.factory';
 import { ITableColumn } from '../../list-table/table.types';
 import { IListConfig, ListViewTypes } from '../../list.component.types';
 import { EndpointCardComponent } from '../endpoint/endpoint-card/endpoint-card.component';
 import { endpointColumns } from '../endpoint/endpoints-list-config.service';
-import { BaseEndpointsDataSource } from './base-endpoints-data-source';
+import { CFEndpointsDataSource } from './cf-endpoints-data-source';
 
 
 @Injectable()
 export class CFEndpointsListConfigService implements IListConfig<EndpointModel> {
   columns: ITableColumn<EndpointModel>[];
   isLocal = true;
-  dataSource: BaseEndpointsDataSource;
+  dataSource: CFEndpointsDataSource;
   viewType = ListViewTypes.CARD_ONLY;
   cardComponent = EndpointCardComponent;
   text = {
@@ -26,12 +29,20 @@ export class CFEndpointsListConfigService implements IListConfig<EndpointModel> 
   tableFixedRowHeight = true;
 
   constructor(
-    private store: Store<AppState>
+    private store: Store<AppState>,
+    paginationMonitorFactory: PaginationMonitorFactory,
+    entityMonitorFactory: EntityMonitorFactory,
+    internalEventMonitorFactory: InternalEventMonitorFactory
   ) {
     this.columns = endpointColumns.filter(column => {
       return column.columnId !== 'type';
     });
-    this.dataSource = new BaseEndpointsDataSource(this.store, this, 'cf');
+    this.dataSource = new CFEndpointsDataSource(
+      this.store,
+      this,
+      paginationMonitorFactory,
+      entityMonitorFactory,
+      internalEventMonitorFactory);
   }
   public getColumns = () => this.columns;
   public getGlobalActions = () => [];
